# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservation Backend API",
        "description": "API documentation for the Apartment Reservation System",
        "contact": {
            "email": "support@reservation.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/apartments": {
            "get": {
                "tags": [
                    "Apartments"
                ],
                "summary": "Get all apartments",
                "description": "Retrieve a paginated list of apartments with optional filters",
                "operationId": "fcdb7dd412537060312e29eb591a8643",
                "parameters": [
                    {
                        "name": "city_id",
                        "in": "query",
                        "description": "Filter by city ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "governorate_id",
                        "in": "query",
                        "description": "Filter by governorate ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price filter",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price filter",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title and description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field (created_at, price, etc.)",
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Sort direction",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apartments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Apartments retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apartments/{id}": {
            "get": {
                "tags": [
                    "Apartments"
                ],
                "summary": "Get apartment by ID",
                "description": "Retrieve detailed information about a specific apartment",
                "operationId": "c9bc439923ea8d32e7109624d47d667b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Apartment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apartment retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Apartment retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Apartment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apartments/favorites": {
            "get": {
                "tags": [
                    "Apartments"
                ],
                "summary": "Get favorite apartments",
                "description": "Retrieve all apartments marked as favorite by the authenticated user",
                "operationId": "30da6fbd280273e6d2334301f51c02f5",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite apartments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Favorite apartments retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apartments/{id}/toggle-favorite": {
            "post": {
                "tags": [
                    "Apartments"
                ],
                "summary": "Toggle favorite status",
                "description": "Add or remove apartment from favorites",
                "operationId": "0f08fd07f56484504ca0be054d3edbf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Apartment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "is_favorite": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Favorite status updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Apartment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apartments/{id}/reviews": {
            "get": {
                "tags": [
                    "Apartments"
                ],
                "summary": "Get apartment reviews",
                "description": "Retrieve all reviews for a specific apartment",
                "operationId": "08a1858f830228c47cefaededf30d743",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Apartment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Reviews retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Apartments"
                ],
                "summary": "Add review to apartment",
                "description": "Submit a review and rating for an apartment",
                "operationId": "a5a8634c7063415ff076712919b66f5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Apartment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "comment"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "example": "Great apartment with amazing view!"
                                    },
                                    "booking_id": {
                                        "description": "Optional booking ID if review is for a completed booking",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Review added successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}